Split out `overlapping_arms`
